pipelines/requirements.txt
kfp==2.7.0
google-cloud-storage
google-cloud-aiplatform
google-cloud-pipeline-components
python-dotenv
